The mcSimulation function now defines the class using S4 options for outputs. Added '@' operator to the as.data.frame.mcSimulation function since the mcSimulation_object is an S4 class.
We added classes to the plot_npv function and have moved the $ to a @ operator for S4 classes (https://www.datamentor.io/r-programming/object-class-introduction/).

We re-documented a bunch of functions since the build option is not doing so.

To do: 1) documentation mcSimulation (add class description to 'value' parameter). 2) vignette in markdown style. 3) how to do it with NPV and cashflow post-hoc. 4) how to make plots more universal ('...' parameter to pass to ggplot2). Not only NPV but also other distribution options.